Add mlockall call to Neovim's main.c
This time in a subtree instead of a submodule so that it actually works.
I also wrote some comments by the lines I changed to be sure I was
confirming with Neovim's Apache license, which states that I need to
"prominently" mark my modifications. I hope this is prominent enough.